Update destructuring 2.md (#63)
* Update destructuring 2.md

I'm very new to Rust and the changes I proposed may either be incorrect or not fit in the flow of the document. Please feel free to suggest modifications that I can make or to reject altogether. Thanks.

* Update destructuring 2.md

Add information about destructors and their effect on allowing `Copy` trait to be implemented by the struct in question.

* Update destructuring 2.md

Sorry about the delay, I was going to blame github but I checked and there indeed was an email notifying me about your review. 

I've addressed your comments: 

1) Mention that all fields of a type must implement `Copy` for that type to implement `Copy`
2) Replace the word struct with types in a few places. (I've left it at places where it was a keyword)
3) Dropped the sentence- "if an object has a destructor then it has move semantics" 

Feel free to let me know if there's still things that I should improve- hopefully I'll see the email in time